requirement.txt
langchain
python-dotenv
langchain-openai
langchain_core
langchain_community
langchain-groq
streamlit
langchain_huggingface
faiss-cpu
pypdf
langchain_chroma
langchain
python-dotenv
ipykernel
langchain-community
pypdf
bs4
arxiv
pymupdf
wikipedia
langchain-text-splitters
langchain-openai
chromadb
sentence_transformers
langchain_huggingface
faiss-cpu
langchain_chroma
duckdb
pandas
openai
langchain-groq
duckduckgo-search
pymupdf
arxiv
wikipedia
mysql-connector-python
SQLAlchemy
validators==0.28.1
youtube_transcript_api
unstructured
pytube
numexpr
huggingface_hub
cassio
datasets
tiktoken
PyPDF2
gradio
newspaper3k
